feat!(config): depr. options in favor of nested border config
Note:
These changes are 'technically' a breaking change since users relying on
the root options for variaous border settings will need to update their
configs.
However, to prevent outright breakage, the old options will still work
for the time being and are internally migrated to the new nested config.
The migration process will be removed in a future major release.